Refactorización de estructura de archivos
