+++ lib/domain/collections/color_by_type.dart
