+++ lib/domain/models/pokemon_model.dart
@@ -1,3 +1,5 @@
+import 'package:mobile_pokedex/domain/models/stat_model.dart';
+
 class Pokemon{
   final int id;
   final String name;
@@ -9,58 +11,31 @@ class Pokemon{
   Pokemon({
     required this.id,
     required this.name,
-    required this.height,
-    required this.weight,
+    required int apiHeight,
+    required int apiWeight,
     required this.types,
     required this.stats
-  });
+  }):
+  height = normalizeHeight(apiHeight), 
+  weight = normalizeWeight(apiWeight);
 
-  Map<String, dynamic> toJSON(){
-    return{
-      'id': id,
-      'name': name,
-      'types': types,
-      'stats':stats
-    };
-  }
-
-  //Probar como usar estas funciones en factory
-  double normalizeHeight(int heightInDecimeters){
+  static double normalizeHeight(int heightInDecimeters){
     return heightInDecimeters / 10;
   }
 
-  double normalizeWeight(int weightInGrams){
-    return weightInGrams / 100;
+  static double normalizeWeight(int weightInGrams){
+    return weightInGrams/10;
   }
 
   factory Pokemon.fromJson(Map<String, dynamic> json){
     return Pokemon(
       id: json['id'], 
       name: json['name'],
-      height: json['height'] / 10,
-      weight: json['weight'] / 10, 
+      apiHeight: json['height'],
+      apiWeight: json['weight'], 
       types: json['types'].map((type)=> PokemonType(name: type['type']['name'])).toList(), 
       stats: json['stats'].map((stat)=> Stat(originalName: stat['stat']['name'], baseValue: stat['base_stat'])).toList()
     );
-  }
-}
-
-class Stat{
-  String name;
-  int baseValue;
-
-  Stat({
-    required String originalName,
-    required this.baseValue
-  }): name = changeStatName(originalName);
-
-  static String changeStatName(String name){
-    final Map<String, String> namesToBeChanged = {
-      "special-attack": "special Attack",
-      "special-defense": "special Defense"
-    }; 
-
-    return namesToBeChanged[name] ?? name;
   }
 }
