+++ lib/presentation/widgets/SecondaryCard/TypeCard/type_card.dart
@@ -2,9 +2,9 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:mobile_pokedex/models/pokemon_model.dart';
+import 'package:mobile_pokedex/domain/models/pokemon_model.dart';
 //collections
-import '../../../collections/color_by_type.dart';
+import '../../../../domain/collections/color_by_type.dart';
 
 class TypeCard extends StatelessWidget{
   TypeCard({
